backend/requirements.txt
# 核心框架
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
greenlet>=3.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# 认证
email-validator>=2.0.0
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.0
python-multipart>=0.0.6

# LLM/Agent
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
langgraph>=0.0.20
openai>=1.0.0

# 工具
httpx>=0.25.0
openpyxl>=3.1.0
